% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citiesPop.R
\docType{data}
\name{citiesPop}
\alias{citiesPop}
\title{Largest Cities Since 1800}
\format{
A data frame of columns 'NAME,' 'YEAR,' 'LAT,' 'LONG,' and 'POP'
}
\usage{
citiesPop
}
\description{
Data set recording the largest cities worldwide from 1800 to 2025, sourced from Wikipedia's article
"Historical urban community sizes." The top ten cities per year are included.
}
\keyword{datasets}
