% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Historic Map}
\usage{
map(country, year, data = historicData, ...)
}
\arguments{
\item{country}{The country that the user wants to visualize}

\item{year}{The year the user wants to see reflected by borders}

\item{data}{The dataset of historical borders}

\item{...}{Included for flexibility}
}
\value{
A map of a historic territory's borders at a given year.
}
\description{
A function to map historic states and borders.
}
\details{
This function will map a historic country (or historic borders of a modern country) included
in the historicalBorders package. Country names must match the names in the dataset,
and if a year is specified that is not in the data set, the function will map the borders
at the closest year on record. Additional data can be passed to visualize over this map, such as population or development data.
}
\examples{
map("USA", 1945)


}
