% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCountry.R
\name{mapCountry}
\alias{mapCountry}
\title{Map Countries}
\usage{
mapCountry(year, country, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{country}{The country a user wishes to map}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
A map of a country at a given time.
}
\description{
Map a certain country's historic borders.
}
\details{
This function will generate a map of any country for any year since 1800 that the user
inputs. If a given year is not in the database, the function will show a map for the closest year via
rounding. A user can also provide additional data to be mapped over these borders, in which case
they must also provide an aesthetic specification for which variable(s) will be mapped and how.
The function will return a ggplot2 object, from which a user can further customize their output.

Note that country's names have changed over time, and referring to an area as it's modern
name will not always return the desired outcome For example, calling mapCountry(1800, "Germany")
will return a map of Germany from 1880, as the modern country of Germany did not exist at the
turn of the 19th century.
}
\examples{
mapCountry(1880, "Germany")
mapCountry(1875, "Zambia")

}
