% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNorthAmerica.R
\name{mapNorthAmerica}
\alias{mapNorthAmerica}
\title{North America Map}
\usage{
mapNorthAmerica(year, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
This function will return a map of North American at a given time period.
}
\description{
Generates a map of North America at a given year since 1800.

#' @details
This function will generate a map of North America for any year since 1800 that the user
inputs. If a given year is not in the database, the function will show a map for the closest year via
rounding. A user can also provide additional data to be mapped over these borders, in which case
they must also provide an aesthetic specification for which variables(s) will be mapped and how.
The function will return a ggplot2 object, from which a user can further customize their output.
}
\examples{
mapNorthAmerica(1800)
mapNorthAmerica(1900, dataset = population1900, aesthetics = aes(fill = POP))

}
