% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOceania.R
\name{mapOceania}
\alias{mapOceania}
\title{Oceania Map}
\usage{
mapOceania(year, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
This function will return a map of Oceania at a given time period.
}
\description{
Generates a map of Oceania at a given year since 1800.
}
\examples{
mapOceania(1825)
mapOceania(1900, dataset = population1900, aesthetics = aes(fill = POP))

}
