% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_regions.R
\name{options_regions}
\alias{options_regions}
\title{Available Regions}
\usage{
options_regions(continent = NULL)
}
\arguments{
\item{continent}{Optional, filters available regions to a single continent}
}
\value{
A data frame of available regions
}
\description{
Available Regions
}
