% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsYears.R
\name{optionsYears}
\alias{optionsYears}
\title{Available Years}
\usage{
optionsYears()
}
\value{
List of years for which maps are available.
}
\description{
Returns all years for which historical maps are available within the historicalBorders package.
If a user tries to visualize a map outside of these options, historicalBorders functions will round
to the closest year given by this function.
}
\examples{
optionsYears()

}
