% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population1900.R
\docType{data}
\name{population1900}
\alias{population1900}
\title{Population of the World at 1900}
\format{
A data frame of columns "NAME" (name of country) and "POP" (population)
}
\usage{
population1900
}
\description{
A data frame storing the population of the world in 1900. Build from the Wikipedia article "List of countries by population in 1900."
Some historic nations are missing from this set.
}
\keyword{datasets}
