% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{world}
\alias{world}
\title{My awesome spatial data}
\format{
An sf data frame with X rows and Y variables:
\describe{
  \item{column1}{Description of column 1}
  \item{geometry}{The sf geometry column}
  ...
}
}
\source{
Where the data comes from
}
\usage{
world
}
\description{
A description of the data.
}
\keyword{datasets}
