% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww1EuropeFatality.R
\docType{data}
\name{ww1EuropeFatality}
\alias{ww1EuropeFatality}
\title{European Fatalities in World War I}
\format{
A data frame of columns 'NAME,' 'DEATHS_TOTAL,' 'DEATHS_MILITARY', and 'DEATHS_CIV'
}
\usage{
ww1EuropeFatality
}
\description{
Data set recording the death toll, both military and civilian, for all European countries involved with World War I.
}
\keyword{datasets}
