% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAfrica.R
\name{mapAfrica}
\alias{mapAfrica}
\title{Africa Map}
\usage{
mapAfrica(year, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
This function will generate a historic map ggplot object of Africa at a given time period.
}
\description{
Generates a map of Africa at a given year since 1800.
}
\examples{
mapAfrica(1800)
mapAfrica(1900, dataset = population1900, aesthetics = aes(fill = POP))
}
