% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAfrica.R
\name{mapAfrica}
\alias{mapAfrica}
\title{Africa Map}
\usage{
mapAfrica(year, dataset = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders}

\item{dataset}{The dataset of historical borders.}
}
\value{
This function will generate a historic map ggplot object of Africa at a given time period.
}
\description{
Africa Map
}
\examples{
mapAfrica(1800)
mapAfrica(1850)
}
