% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAsia.R
\name{mapAsia}
\alias{mapAsia}
\title{Asia Map}
\usage{
mapAsia(year, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{The dataset of historical borders.}
}
\value{
This function will return a map of Asia at a given time period.
}
\description{
Generates a map of Asia at a given year since 1800.
}
\examples{
mapAsia(1820)
mapAsia(1890)

}
