% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCustom.R
\name{mapCustom}
\alias{mapCustom}
\title{Custom Map}
\usage{
mapCustom(year, lat, long, dataset = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{lat}{The latitude of the area the user wants to map.}

\item{long}{The longitude of the area the user wants to map.}

\item{dataset}{The dataset of historical borders.}
}
\value{
A custom map of the latitude and longitudes specified by R.
}
\description{
Custom Map
}
\examples{
mapCustom(1800, 43.29N, 74.21W)
mapCustom(1820, 48.85N, 2.35E)

}
