% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCustom.R
\name{mapCustom}
\alias{mapCustom}
\title{Custom Map}
\usage{
mapCustom(year, lat, long, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{lat}{The latitude of the area the user wants to map.}

\item{long}{The longitude of the area the user wants to map.}

\item{dataset}{The dataset of historical borders.}
}
\value{
A custom map of the latitude and longitudes specified by R.
}
\description{
Generates a map of a custom area at a given year since 1800. Area to be mapped can be specified
using coordinate pairs to form a rectangle around a certain area of the globe.
}
\examples{
mapCustom(1800, c(43.29, 57.10), c(74.21, 100.7))
mapCustom(1820, c(48.85, 55.5), c(2.35, 10.8))

}
