% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCustom.R
\name{mapCustom}
\alias{mapCustom}
\title{Custom Map}
\usage{
mapCustom(year, lat, long, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{lat}{The latitude of the area the user wants to map.}

\item{long}{The longitude of the area the user wants to map.}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
A custom map of the latitude and longitudes specified by R.
}
\description{
Generates a map of a custom area at a given year since 1800. Area to be mapped can be specified
using coordinate pairs to form a rectangle around a certain area of the globe.
}
\details{
This function will generate a map of a custom area for any year since 1800 that the user
inputs. The user can specify a rectangular area to map by inputting latitude and longitude coordinates.
If a given year is not in the database, the function will show a map for the closest year via
rounding. A user can also provide additional data to be mapped over these borders, in which case
they must also provide an aesthetic specification for which variables(s) will be mapped and how.
The function will return a ggplot2 object, from which a user can further customize their output.
}
\examples{
mapCustom(1800, c(19.5, 45.4), c(25.3, 41.2))
mapCustom(1900, c(62.7, 130.6), c(-3.3, 44.2), population1900, aes(fill = POP))

}
