% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCustom.R
\name{mapCustom}
\alias{mapCustom}
\title{Custom Map}
\usage{
mapCustom(year, lat, long, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{lat}{The latitude of the area the user wants to map.}

\item{long}{The longitude of the area the user wants to map.}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
A custom map of the latitude and longitudes specified by R.
}
\description{
Generates a map of a custom area at a given year since 1800. Area to be mapped can be specified
using coordinate pairs to form a rectangle around a certain area of the globe.
}
\examples{
mapCustom(1800, c(19.5, 45.4), c(25.3, 41.2))
mapCustom(1900, c(62.7, 130.6), c(-3.3, 44.2), dataset = population1900, aesthetics = aes(fill = POP))

}
