% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRegion.R
\name{mapRegion}
\alias{mapRegion}
\title{mapRegion}
\usage{
mapRegion(year, region, dataset = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{region}{The region a user wants to map}

\item{dataset}{The dataset of historical borders.}
}
\value{
This function will return a map of a region at a given time period.
}
\description{
mapRegion
}
\examples{
mapRegion(1825, "Eastern Europe")
mapRegion(1875, "Western Africa")
}
