% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRegion.R
\name{mapRegion}
\alias{mapRegion}
\title{Map Historic Regions}
\usage{
mapRegion(year, region, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{region}{The region a user wants to map}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
A region at a given time period.
}
\description{
Maps a region of the world using historic borders.
}
\details{
This function will generate a map of a region of the world for any year since 1800. If a given year is not in the database, the function will show a map for the closest year via
rounding. A user can also provide additional data to be mapped over these borders, in which case
they must also provide an aesthetic specification for which variables(s) will be mapped and how.
The function will return a ggplot2 object, from which a user can further customize their output.
Regions are based on the United Nations geoscheme, and options can be viewed using the optionsRegions function.
}
\examples{
mapRegion(1914, region = "Western Europe", ww1EuropeFatality, aes(fill = DEATHS_TOTAL))
mapRegion(1875, "Western Africa")

}
