% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRegion.R
\name{mapRegion}
\alias{mapRegion}
\title{Map Historic Regions}
\usage{
mapRegion(year, region, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{region}{The region a user wants to map}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
A region at a given time period.
}
\description{
Maps a region of the world using historic borders.
}
\details{
Available regions to map can be viewed using optionsRegions.
}
\examples{
mapRegion(1914, region = "Western Europe", dataset = ww1EuropeFatality, aesthetics = aes(fill = DEATHS_TOTAL))
mapRegion(1875, "Western Africa")

}
