% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSouthAmerica.R
\name{mapSouthAmerica}
\alias{mapSouthAmerica}
\title{South America Map}
\usage{
mapSouthAmerica(year, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
This function will return a map of South America at a given time period.
}
\description{
Generates a map of South America at a given year since 1800.
}
\examples{
mapSouthAmerica(1827)
mapSouthAmerica(1900, dataset = population1900, aesthetics = aes(fill = POP))
}
