% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSouthAmerica.R
\name{mapSouthAmerica}
\alias{mapSouthAmerica}
\title{South America Map}
\usage{
mapSouthAmerica(year, dataset = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{The dataset of historical borders.}
}
\value{
This function will return a map of South America at a given time period.
}
\description{
South America Map
}
\examples{
mapSouthAmerica(1800)
mapSouthAmerica(1827)
}
