% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapworld.R
\name{mapworld}
\alias{mapworld}
\title{World Map}
\usage{
mapworld(year, dataset = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{The optional additional dataset to graph onto a map}
}
\value{
A world map ggplot object, with borders dated to a given year
}
\description{
Map the world with historic borders.
}
\details{
This function will generate a historic map of the world at a given point in time.
Users can input additional data sets that they wish to map onto the historic map.
If users try to map a year that is not in the existing data, the world map closest to
that year will be mapped instead.
}
\examples{
mapworld(1800)
mapworld(1875)

}
