% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapWorld.R
\name{mapWorld}
\alias{mapWorld}
\title{World Map}
\usage{
mapWorld(year, dataset = NULL, aesthetics = NULL)
}
\arguments{
\item{year}{The year the user wants to see reflected by borders.}

\item{dataset}{Additional data to map over historical borders.}

\item{aesthetics}{The aesthetic features (ex. fill, color) to map onto the borders}
}
\value{
A world map ggplot object, with borders dated to a given year
}
\description{
Map the world with historic borders.
}
\details{
This function will generate a historic map of the world at a given point in time.
Users can input additional data sets that they wish to map onto the historic map.
If users try to map a year that is not in the existing data, the world map closest to
that year will be mapped instead.
}
\examples{
mapWorld(1800)
mapWorld(1900, dataset = population1900, aesthetics = aes(fill = POP))

}
