% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsRegions.R
\name{optionsRegions}
\alias{optionsRegions}
\title{Available Regions}
\usage{
optionsRegions(continent = NULL)
}
\arguments{
\item{continent}{Optional, filters available regions to a single continent}
}
\value{
A data frame of available regions
}
\description{
Prints regions that can be mapped using mapRegion(), along with the modern countries within these regions.
Regions are defined using the United Nations' geoscheme designation.
}
\examples{
optionsRegions()
optionsRegions("Africa")

}
