% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions of the World}
\format{
A data frame of columns 'continent,' 'region,' and 'modern_countries'
}
\usage{
regions
}
\description{
Data frame explaining the regions of the world, to be called in `mapRegion()`.
The data contains the regions of all continents, as defined by the UN, along with
which modern-day countries are held within these regions.
}
\keyword{datasets}
