% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\docType{data}
\name{world}
\alias{world}
\title{World data}
\format{
A collection of shapefiles
}
\source{
https://github.com/aourednik/historical-basemaps
}
\usage{
world
}
\description{
Global shape file data across all years
}
\keyword{datasets}
